+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/web/StockMoveController.java
@@ -17,30 +17,16 @@
  */
 package com.axelor.apps.supplychain.web;
 
-import com.axelor.apps.base.db.AppSupplychain;
-import com.axelor.apps.base.db.repo.AppSupplychainRepository;
-import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.apps.stock.db.StockMove;
-import com.axelor.apps.stock.db.StockMoveLine;
-import com.axelor.apps.stock.service.StockMoveLineService;
-import com.axelor.apps.supplychain.exception.IExceptionMessage;
 import com.axelor.apps.supplychain.service.StockMoveServiceSupplychain;
 import com.axelor.exception.service.TraceBackService;
-import com.axelor.i18n.I18n;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
-import com.google.common.base.Strings;
 import com.google.inject.Inject;
-import java.math.BigDecimal;
-import java.util.StringJoiner;
 
 public class StockMoveController {
 
   @Inject private StockMoveServiceSupplychain stockMoveService;
-
-  @Inject private AppSupplychainRepository appSupplychainRepo;
-
-  @Inject private StockMoveLineService stockMoveLineService;
 
   public void addSubLines(ActionRequest request, ActionResponse response) {
     try {
@@ -65,38 +51,12 @@ public class StockMoveController {
   }
 
   public void verifyProductStock(ActionRequest request, ActionResponse response) {
-    StockMove stockMove = request.getContext().asType(StockMove.class);
-    AppSupplychain appSupplychain = appSupplychainRepo.all().fetchOne();
-    StringJoiner notAvailableProducts = new StringJoiner(",");
-    if (stockMove.getAvailabilityRequest()
-        && stockMove.getStockMoveLineList() != null
-        && appSupplychain.getIsVerifyProductStock()
-        && stockMove.getFromStockLocation() != null) {
-
-      int counter = 1;
-      for (StockMoveLine stockMoveLine : stockMove.getStockMoveLineList()) {
-        SaleOrderLine saleOrderLine = stockMoveLine.getSaleOrderLine();
-        if (saleOrderLine != null) {
-          stockMoveLineService.updateAvailableQty(stockMoveLine, stockMove.getFromStockLocation());
-          BigDecimal availableQty =
-              stockMoveLine.getTrackingNumber() != null
-                  ? stockMoveLine.getAvailableQty()
-                  : stockMoveLine.getAvailableQtyForProduct();
-          BigDecimal realQty = stockMoveLine.getRealQty();
-          if (availableQty.compareTo(realQty.subtract(saleOrderLine.getReservedQty())) < 0
-              && counter <= 10) {
-            notAvailableProducts.add(stockMoveLine.getProduct().getFullName());
-            counter++;
-          }
-        }
-      }
-      if (!Strings.isNullOrEmpty(notAvailableProducts.toString())) {
-        response.setValue("availabilityRequest", false);
-        response.setFlash(
-            String.format(
-                I18n.get(IExceptionMessage.STOCK_MOVE_VERIFY_PRODUCT_STOCK_ERROR),
-                notAvailableProducts.toString()));
-      }
+    try {
+      StockMove stockMove = request.getContext().asType(StockMove.class);
+      stockMoveService.verifyProductStock(stockMove);
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+      response.setReload(true);
     }
   }
 }